app/requirements.txt
Django>=4.0.10,<4.1
djangorestframework>=3.14.0,<3.15
psycopg2>=2.9.5,<2.10
drf-spectacular>=0.26.0,<0.27
django-bitfield>=2.2.0,<2.3
Pillow>=9.4.0,<9.5
uwsgi>=2.0.21,<2.1